Split enterprise delegation validation
